Move TextAnnotate to outer package.
