+++ codeviz_backend/src/main/java/codeViz/gitHistory/CommitInfo.java
@@ -1,5 +1,6 @@
 package codeViz.gitHistory;
 
+import codeViz.TextAnnotate;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.StringReader;
